attack-library-system/requirements.txt
# Core dependencies for Attack Library System
pydantic>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
click>=8.1.0
pyyaml>=6.0.1
jinja2>=3.1.2
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
aiohttp>=3.9.0
rich>=13.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0
jsonschema>=4.17.0
faker>=19.0.0
nltk>=3.8.0
textstat>=0.7.3

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Optional ML dependencies
transformers>=4.33.0
torch>=2.0.0
scikit-learn>=1.3.0
